once/ installer updates.

+++ .fries/once/vendor_dubsacks.sh
@@ -1,9 +1,11 @@
 # File: vendor_dubsacks.sh
 # Author: Landon Bouma (landonb &#x40; retrosoft &#x2E; com)
-# Last Modified: 2016.11.03
+# Last Modified: 2016.11.12
 # Project Page: https://github.com/landonb/home_fries
 # Summary: Dubsacks VIM setup script.
 # License: GPLv3
+
+set -e
 
 if [[ -z ${DO_INSTALL_DUBSACKS+x} ]]; then
   DO_INSTALL_DUBSACKS=true
@@ -19,46 +21,48 @@ stage_4_dubsacks_install () {
 
   pushd ${HOME} &> /dev/null
 
-  if [[ ! -h ${HOME}/.vimrc ]]; then
+  if [[ -e ${HOME}/.vimrc && ! -h ${HOME}/.vimrc ]]; then
     echo
-    echo "WARNING: Was expecting .vimrc to be a symlink: moving it aside: ${HOME}/.vimrc"
+    echo "WARNING: Was not expecting ~/.vimrc: moving it aside"
     mv ${HOME}/.vimrc ${HOME}/BACKUP-vimrc-`date +%Y_%m_%d`-`uuidgen`
-    /bin/ln -s .vim/bundle/dubs_all/.vimrc.bundle .vimrc
   fi
 
   if [[ -e ${HOME}/.vimprojects ]]; then
     echo
-    echo "WARNING: Not expecting vimprojects: moving it aside: ${HOME}/.vimprojects"
-    mv ${HOME}/.vimprojects ${HOME}/BACKUP-vimprojects-`date +%Y_%m_%d`-`uuidgen`
+    echo "WARNING: Was not expecting ~/.vimprojects: moving it aside"
+    mv ${HOME}/.vimprojects ${HOME}/.vimprojects-TBD-`date +%Y_%m_%d`-`uuidgen`
   fi
 
-  if [[ -e ${HOME}/.vim && ! -d ${HOME}/.vim/.git ]]; then
-    echo
-    echo "WARNING: Was expecting vim to be a git repo: moving it aside: ${HOME}/.vim"
-    mv ${HOME}/.vim ${HOME}/BACKUP-vim-`date +%Y_%m_%d`-`uuidgen`
+  if [[ -e ${HOME}/.vim ]]; then
+    if [[ ! -d ${HOME}/.vim/.git ]]; then
+      echo
+      echo "WARNING: Was expecting ~/.vim to be a git repo: moving it aside"
+      mv ${HOME}/.vim ${HOME}/.vim-TBD-`date +%Y_%m_%d`-`uuidgen`
+    else 
+      cd ${HOME}/.vim &> /dev/null
+      git +e
+      git remote -v | grep "landonb/dubsacks_vim.git" > /dev/null
+      exit_code=$?
+      git -e
+      if [[ $exit_code -ne 0 ]]; then
+        echo
+        echo "WARNING: Was expecting ~/.vim to be the Dubsacks repo: moving it aside"
+        mv ${HOME}/.vim ${HOME}/.vim-TBD-`date +%Y_%m_%d`-`uuidgen`
+      fi
+    fi
   fi
+
   if [[ ! -e ${HOME}/.vim ]]; then
     git clone ${URI_DUBSACKS_VIM_GIT} ${HOME}/.vim
-
-    # The clone doesn't grab the submodules.
-    # Also, create the ~/.vimrc symlink.
-    pushd ${HOME}/.vim &> /dev/null
-    ./setup.sh
-    popd &> /dev/null
+    cd ${HOME}/.vim
   else
-    pushd ${HOME}/.vim &> /dev/null
-    # We'll just assume that *we* setup .vim already,
-    # i.e., that this is our repo and not something else.
-    # We could check that, say, ~/.vim/bundle/dubs_core exists,
-    # but it's unlikely that that'll not be there, and if it isn't,
-    # then you got bigger problems.
+    cd ${HOME}/.vim
     git pull
-
-    # Update the submodules.
-    git submodule update --init --remote
-
-    popd &> /dev/null
   fi
+
+  # Grab or update the submodules.
+  # Also, create the ~/.vimrc symlink.
+  ./setup.sh
 
   popd &> /dev/null
 
@@ -66,17 +70,6 @@ stage_4_dubsacks_install () {
   source custom_setup.extras.sh
   stage_4_font_typeface_hack
   stage_4_parT_install
-
-  if false; then
-    # FIXME/MAYBE: Implement this:
-    pushd ${HOME} &> /dev/null
-    m4 \
-      --define=YOUR_FULL_NAME_HERE=${YOUR_FULL_NAME_HERE} \
-      --define=YOUR_EMAIL_ADDY_HERE=${YOUR_EMAIL_ADDY_HERE} \
-      --define=YOUR_GITHUB_USERNAME=${YOUR_GITHUB_USERNAME} \
-      .cookiecutterrc.m4
-    popd &> /dev/null
-  fi
 
 } # end: stage_4_dubsacks_install
 
@@ -96,6 +89,6 @@ if [[ "$0" == "$BASH_SOURCE" ]]; then
   setup_dubsacks_go
 fi
 
-# Stupid message for debugging ./setup_linux.sh
+# Stupid message for debugging ./setup_ubuntu.sh
 echo 'Thank you for dubssacking!'
